chore: Fix tests
