+++ packages/react-native-reanimated/plugin/index.js
@@ -1,3 +1,3 @@
-"use strict";var Ke=Object.create;var me=Object.defineProperty;var Qe=Object.getOwnPropertyDescriptor;var et=Object.getOwnPropertyNames;var tt=Object.getPrototypeOf,rt=Object.prototype.hasOwnProperty;var f=(e,t)=>()=>(t||e((t={exports:{}}).exports,t),t.exports);var nt=(e,t,r,n)=>{if(t&&typeof t=="object"||typeof t=="function")for(let i of et(t))!rt.call(e,i)&&i!==r&&me(e,i,{get:()=>t[i],enumerable:!(n=Qe(t,i))||n.enumerable});return e};var P=(e,t,r)=>(r=e!=null?Ke(tt(e)):{},nt(t||!e||!e.__esModule?me(r,"default",{value:e,enumerable:!0}):r,e));var E=f(m=>{"use strict";Object.defineProperty(m,"__esModule",{value:!0});m.workletClassFactorySuffix=m.isWorkletizableObjectNode=m.isWorkletizableObjectPath=m.isWorkletizableFunctionNode=m.isWorkletizableFunctionPath=m.WorkletizableObject=m.WorkletizableFunction=void 0;var B=require("@babel/types");m.WorkletizableFunction="FunctionDeclaration|FunctionExpression|ArrowFunctionExpression|ObjectMethod";m.WorkletizableObject="ObjectExpression";function it(e){return e.isFunctionDeclaration()||e.isFunctionExpression()||e.isArrowFunctionExpression()||e.isObjectMethod()}m.isWorkletizableFunctionPath=it;function ot(e){return(0,B.isFunctionDeclaration)(e)||(0,B.isFunctionExpression)(e)||(0,B.isArrowFunctionExpression)(e)||(0,B.isObjectMethod)(e)}m.isWorkletizableFunctionNode=ot;function st(e){return e.isObjectExpression()}m.isWorkletizableObjectPath=st;function at(e){return(0,B.isObjectExpression)(e)}m.isWorkletizableObjectNode=at;m.workletClassFactorySuffix="__classFactory"});var V=f(w=>{"use strict";Object.defineProperty(w,"__esModule",{value:!0});w.initializeGlobals=w.globals=w.defaultGlobals=void 0;var lt=["globalThis","Infinity","NaN","undefined","eval","isFinite","isNaN","parseFloat","parseInt","decodeURI","decodeURIComponent","encodeURI","encodeURIComponent","escape","unescape","Object","Function","Boolean","Symbol","Error","AggregateError","EvalError","RangeError","ReferenceError","SyntaxError","TypeError","URIError","InternalError","Number","BigInt","Math","Date","String","RegExp","Array","Int8Array","Uint8Array","Uint8ClampedArray","Int16Array","Uint16Array","Int32Array","Uint32Array","BigInt64Array","BigUint64Array","Float32Array","Float64Array","Map","Set","WeakMap","WeakSet","ArrayBuffer","SharedArrayBuffer","DataView","Atomics","JSON","WeakRef","FinalizationRegistry","Iterator","AsyncIterator","Promise","GeneratorFunction","AsyncGeneratorFunction","Generator","AsyncGenerator","AsyncFunction","Reflect","Proxy","Intl","null","this","global","window","globalThis","console","performance","queueMicrotask","requestAnimationFrame","setImmediate","arguments","HermesInternal","_WORKLET","ReanimatedError","__reanimatedLoggerConfig"],ct=["_IS_FABRIC","_log","_toString","_scheduleHostFunctionOnJS","_scheduleRemoteFunctionOnJS","_scheduleOnRuntime","_makeShareableClone","_updatePropsPaper","_updatePropsFabric","_removeFromPropsRegistry","_measurePaper","_measureFabric","_scrollToPaper","_dispatchCommandPaper","_dispatchCommandFabric","_setGestureState","_notifyAboutProgress","_notifyAboutEnd","_runOnUIQueue","_getAnimationTimestamp"];w.defaultGlobals=new Set(lt.concat(ct));function ut(){w.globals=new Set(w.defaultGlobals)}w.initializeGlobals=ut});var D=f(W=>{"use strict";Object.defineProperty(W,"__esModule",{value:!0});W.replaceWithFactoryCall=W.addCustomGlobals=W.isRelease=void 0;var G=require("@babel/types"),dt=V();function ft(){var e,t;let r=/(prod|release|stag[ei])/i;return!!(!((e=process.env.BABEL_ENV)===null||e===void 0)&&e.match(r)||!((t=process.env.NODE_ENV)===null||t===void 0)&&t.match(r))}W.isRelease=ft;function pt(){this.opts&&Array.isArray(this.opts.globals)&&this.opts.globals.forEach(e=>{dt.globals.add(e)})}W.addCustomGlobals=pt;function bt(e,t,r){if(!t||!mt(e))e.replaceWith(r);else{let n=(0,G.variableDeclaration)("const",[(0,G.variableDeclarator)((0,G.identifier)(t),r)]);e.replaceWith(n)}}W.replaceWithFactoryCall=bt;function mt(e){return(0,G.isScopable)(e.parent)||(0,G.isExportNamedDeclaration)(e.parent)}});var J=f(M=>{"use strict";var yt=M&&M.__rest||function(e,t){var r={};for(var n in e)Object.prototype.hasOwnProperty.call(e,n)&&t.indexOf(n)<0&&(r[n]=e[n]);if(e!=null&&typeof Object.getOwnPropertySymbols=="function")for(var i=0,n=Object.getOwnPropertySymbols(e);i<n.length;i++)t.indexOf(n[i])<0&&Object.prototype.propertyIsEnumerable.call(e,n[i])&&(r[n[i]]=e[n[i]]);return r};Object.defineProperty(M,"__esModule",{value:!0});M.workletTransformSync=void 0;var kt=require("@babel/core");function gt(e,t){let{extraPlugins:r=[],extraPresets:n=[]}=t,i=yt(t,["extraPlugins","extraPresets"]);return(0,kt.transformSync)(e,Object.assign(Object.assign({},i),{plugins:[...Ot,...r],presets:[..._t,...n]}))}M.workletTransformSync=gt;var _t=[require.resolve("@babel/preset-typescript")],Ot=[require.resolve("@babel/plugin-transform-shorthand-properties"),require.resolve("@babel/plugin-transform-arrow-functions"),require.resolve("@babel/plugin-transform-optional-chaining"),require.resolve("@babel/plugin-transform-nullish-coalescing-operator"),[require.resolve("@babel/plugin-transform-template-literals"),{loose:!0}]]});var ge=f(O=>{"use strict";var vt=O&&O.__createBinding||(Object.create?function(e,t,r,n){n===void 0&&(n=r);var i=Object.getOwnPropertyDescriptor(t,r);(!i||("get"in i?!t.__esModule:i.writable||i.configurable))&&(i={enumerable:!0,get:function(){return t[r]}}),Object.defineProperty(e,n,i)}:function(e,t,r,n){n===void 0&&(n=r),e[n]=t[r]}),ht=O&&O.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),ye=O&&O.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(e!=null)for(var r in e)r!=="default"&&Object.prototype.hasOwnProperty.call(e,r)&&vt(t,e,r);return ht(t,e),t},Et=O&&O.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(O,"__esModule",{value:!0});O.buildWorkletString=void 0;var ke=require("@babel/core"),St=Et(require("@babel/generator")),c=require("@babel/types"),H=require("assert"),It=ye(require("convert-source-map")),xt=ye(require("fs")),wt=E(),Wt=D(),Ft=J(),Ct="mock source map";function jt(e,t,r,n,i){Dt(e,n);let s=e.program.body.find(g=>(0,c.isFunctionDeclaration)(g))||e.program.body.find(g=>(0,c.isExpressionStatement)(g))||void 0;(0,H.strict)(s,"[Reanimated] `draftExpression` is undefined.");let l=(0,c.isFunctionDeclaration)(s)?s:s.expression;(0,H.strict)("params"in l,"'params' property is undefined in 'expression'"),(0,H.strict)((0,c.isBlockStatement)(l.body),"[Reanimated] `expression.body` is not a `BlockStatement`");let u=new Set;(0,ke.traverse)(e,{NewExpression(g){if(!(0,c.isIdentifier)(g.node.callee))return;let x=g.node.callee.name;if(!r.some(Z=>Z.name===x)||u.has(x))return;let A=r.findIndex(Z=>Z.name===x);r.splice(A,1);let oe=x+wt.workletClassFactorySuffix;r.push((0,c.identifier)(oe)),(0,c.assertBlockStatement)(l.body),l.body.body.unshift((0,c.variableDeclaration)("const",[(0,c.variableDeclarator)((0,c.identifier)(x),(0,c.callExpression)((0,c.identifier)(oe),[]))])),u.add(x)}});let p=(0,c.functionExpression)((0,c.identifier)(n),l.params,l.body,l.generator,l.async),k=(0,St.default)(p).code;(0,H.strict)(i,"[Reanimated] `inputMap` is undefined.");let _=!((0,Wt.isRelease)()||t.opts.disableSourceMaps);if(_){i.sourcesContent=[];for(let g of i.sources)i.sourcesContent.push(xt.readFileSync(g).toString("utf-8"))}let R=(0,Ft.workletTransformSync)(k,{extraPlugins:[Mt(r)],compact:!0,sourceMaps:_,inputSourceMap:i,ast:!1,babelrc:!1,configFile:!1,comments:!1});(0,H.strict)(R,"[Reanimated] `transformed` is null.");let j;return _&&(Rt()?j=Ct:(j=It.fromObject(R.map).toObject(),delete j.sourcesContent)),[R.code,JSON.stringify(j)]}O.buildWorkletString=jt;function Dt(e,t){(0,ke.traverse)(e,{FunctionExpression(r){if(!r.node.id){r.stop();return}let n=r.node.id.name;r.scope.rename(n,t)}})}function Rt(){return process.env.REANIMATED_JEST_SHOULD_MOCK_SOURCE_MAP==="1"}function At(e,t,r){t.length===0||!(0,c.isProgram)(e.parent)||(0,c.isExpression)(e.node.body)||e.node.body.body.unshift(r)}function Pt(e){var t;(0,c.isProgram)(e.parent)&&!(0,c.isArrowFunctionExpression)(e.node)&&!(0,c.isObjectMethod)(e.node)&&e.node.id&&e.scope.parent&&((t=e.scope.parent.bindings[e.node.id.name])===null||t===void 0?void 0:t.references)>0&&e.node.body.body.unshift((0,c.variableDeclaration)("const",[(0,c.variableDeclarator)((0,c.identifier)(e.node.id.name),(0,c.memberExpression)((0,c.thisExpression)(),(0,c.identifier)("_recur")))]))}function Mt(e){let t=(0,c.variableDeclaration)("const",[(0,c.variableDeclarator)((0,c.objectPattern)(e.map(r=>(0,c.objectProperty)((0,c.identifier)(r.name),(0,c.identifier)(r.name),!1,!0))),(0,c.memberExpression)((0,c.thisExpression)(),(0,c.identifier)("__closure")))]);return{visitor:{"FunctionDeclaration|FunctionExpression|ArrowFunctionExpression|ObjectMethod":r=>{At(r,e,t),Pt(r)}}}}});var ve=f(q=>{"use strict";var qt=q&&q.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(q,"__esModule",{value:!0});q.makeWorkletFactory=void 0;var Lt=require("@babel/core"),zt=qt(require("@babel/generator")),o=require("@babel/types"),F=require("assert"),Oe=require("path"),Tt=V(),_e=E(),le=D(),Nt=ge(),Ut=J(),Bt=require("../package.json").version,Gt="x.y.z";function Ht(e,t){Zt(e),(0,F.strict)(t.file.opts.filename,"[Reanimated] `state.file.opts.filename` is undefined.");let r=(0,zt.default)(e.node,{sourceMaps:!0,sourceFileName:t.file.opts.filename});r.code="("+(e.isObjectMethod()?"function ":"")+r.code+`
-)`;let n=(0,Ut.workletTransformSync)(r.code,{filename:t.file.opts.filename,ast:!0,babelrc:!1,configFile:!1,inputSourceMap:r.map});(0,F.strict)(n,"[Reanimated] `transformed` is undefined."),(0,F.strict)(n.ast,"[Reanimated] `transformed.ast` is undefined.");let i=$t(n.ast,e),s=(0,o.cloneNode)(e.node),l=(0,o.isBlockStatement)(s.body)?(0,o.functionExpression)(null,s.params,s.body,s.generator,s.async):s,{workletName:u,reactName:p}=Xt(e,t),[k,_]=(0,Nt.buildWorkletString)(n.ast,t,i,u,n.map);(0,F.strict)(k,"[Reanimated] `funString` is undefined.");let R=Jt(k),j=1;i.length>0&&(j-=i.length+2);let g=e.parentPath.isProgram()?e:e.findParent(y=>{var se,ae;return(ae=(se=y.parentPath)===null||se===void 0?void 0:se.isProgram())!==null&&ae!==void 0?ae:!1});(0,F.strict)(g,"[Reanimated] `pathForStringDefinitions` is null."),(0,F.strict)(g.parentPath,"[Reanimated] `pathForStringDefinitions.parentPath` is null.");let x=g.parentPath.scope.generateUidIdentifier(`worklet_${R}_init_data`),A=(0,o.objectExpression)([(0,o.objectProperty)((0,o.identifier)("code"),(0,o.stringLiteral)(k))]);if(!(0,le.isRelease)()){let y=t.file.opts.filename;t.opts.relativeSourceLocation&&(y=(0,Oe.relative)(t.cwd,y),_=_==null?void 0:_.replace(t.file.opts.filename,y)),A.properties.push((0,o.objectProperty)((0,o.identifier)("location"),(0,o.stringLiteral)(y)))}_&&A.properties.push((0,o.objectProperty)((0,o.identifier)("sourceMap"),(0,o.stringLiteral)(_))),!(0,le.isRelease)()&&A.properties.push((0,o.objectProperty)((0,o.identifier)("version"),(0,o.stringLiteral)(Vt()?Gt:Bt)));let be=!t.opts.omitNativeOnlyData;be&&g.insertBefore((0,o.variableDeclaration)("const",[(0,o.variableDeclarator)(x,A)])),(0,F.strict)(!(0,o.isFunctionDeclaration)(l),"[Reanimated] `funExpression` is a `FunctionDeclaration`."),(0,F.strict)(!(0,o.isObjectMethod)(l),"[Reanimated] `funExpression` is an `ObjectMethod`.");let U=[(0,o.variableDeclaration)("const",[(0,o.variableDeclarator)((0,o.identifier)(p),l)]),(0,o.expressionStatement)((0,o.assignmentExpression)("=",(0,o.memberExpression)((0,o.identifier)(p),(0,o.identifier)("__closure"),!1),(0,o.objectExpression)(i.map(y=>y.name.endsWith(_e.workletClassFactorySuffix)?(0,o.objectProperty)((0,o.identifier)(y.name),(0,o.memberExpression)((0,o.identifier)(y.name.slice(0,y.name.length-_e.workletClassFactorySuffix.length)),(0,o.identifier)(y.name))):(0,o.objectProperty)((0,o.identifier)(y.name),y,!1,!0))))),(0,o.expressionStatement)((0,o.assignmentExpression)("=",(0,o.memberExpression)((0,o.identifier)(p),(0,o.identifier)("__workletHash"),!1),(0,o.numericLiteral)(R)))];return be&&U.push((0,o.expressionStatement)((0,o.assignmentExpression)("=",(0,o.memberExpression)((0,o.identifier)(p),(0,o.identifier)("__initData"),!1),x))),(0,le.isRelease)()||(U.unshift((0,o.variableDeclaration)("const",[(0,o.variableDeclarator)((0,o.identifier)("_e"),(0,o.arrayExpression)([(0,o.newExpression)((0,o.memberExpression)((0,o.identifier)("global"),(0,o.identifier)("Error")),[]),(0,o.numericLiteral)(j),(0,o.numericLiteral)(-27)]))])),U.push((0,o.expressionStatement)((0,o.assignmentExpression)("=",(0,o.memberExpression)((0,o.identifier)(p),(0,o.identifier)("__stackDetails"),!1),(0,o.identifier)("_e"))))),U.push((0,o.returnStatement)((0,o.identifier)(p))),(0,o.functionExpression)(void 0,[],(0,o.blockStatement)(U))}q.makeWorkletFactory=Ht;function Zt(e){e.traverse({DirectiveLiteral(t){t.node.value==="worklet"&&t.getFunctionParent()===e&&t.parentPath.remove()}})}function Vt(){return process.env.REANIMATED_JEST_SHOULD_MOCK_VERSION==="1"}function Jt(e){let t=e.length,r=5381,n=52711;for(;t--;){let i=e.charCodeAt(t);r=r*33^i,n=n*33^i}return(r>>>0)*4096+(n>>>0)}function Xt(e,t){let r="unknownFile";if(t.file.opts.filename){let l=t.file.opts.filename;r=(0,Oe.basename)(l);let u=l.split("/"),p=u.indexOf("node_modules");p!==-1&&(r=`${u[p+1]}_${r}`)}let n=`${r}${t.workletNumber++}`,i="";(0,o.isObjectMethod)(e.node)&&(0,o.isIdentifier)(e.node.key)?i=e.node.key.name:((0,o.isFunctionDeclaration)(e.node)||(0,o.isFunctionExpression)(e.node))&&(0,o.isIdentifier)(e.node.id)&&(i=e.node.id.name);let s=i?(0,o.toIdentifier)(`${i}_${n}`):(0,o.toIdentifier)(n);return i=i||(0,o.toIdentifier)(n),{workletName:s,reactName:i}}function $t(e,t){let r=new Map,n=new Map;return(0,Lt.traverse)(e,{Identifier(i){if(!i.isReferencedIdentifier())return;let s=i.node.name;if(Tt.globals.has(s)||"id"in t.node&&t.node.id&&t.node.id.name===s)return;let l=i.parent;if((0,o.isMemberExpression)(l)&&l.property===i.node&&!l.computed||(0,o.isObjectProperty)(l)&&(0,o.isObjectExpression)(i.parentPath.parent)&&i.node!==l.value)return;let u=i.scope;for(;u!=null;){if(u.bindings[s]!=null)return;u=u.parent}r.set(s,i.node),n.set(s,!1)}}),t.traverse({Identifier(i){if(!i.isReferencedIdentifier())return;let s=r.get(i.node.name);!s||n.get(i.node.name)||(s.loc=i.node.loc,n.set(i.node.name,!0))}}),Array.from(r.values())}});var he=f(X=>{"use strict";Object.defineProperty(X,"__esModule",{value:!0});X.makeWorkletFactoryCall=void 0;var Yt=require("@babel/types"),Kt=ve();function Qt(e,t){let r=(0,Kt.makeWorkletFactory)(e,t),n=(0,Yt.callExpression)(r,[]);return er(e,n),n}X.makeWorkletFactoryCall=Qt;function er(e,t){let r=e.node.loc;r&&(t.callee.loc={filename:r.filename,identifierName:r.identifierName,start:r.start,end:r.start})}});var $=f(C=>{"use strict";Object.defineProperty(C,"__esModule",{value:!0});C.substituteObjectMethodWithObjectProperty=C.processWorklet=C.processIfWithWorkletDirective=void 0;var ce=require("@babel/types"),tr=E(),rr=D(),nr=he();function Ee(e,t){return!(0,ce.isBlockStatement)(e.node.body)||!ir(e.node.body.directives)?!1:(Se(e,t),!0)}C.processIfWithWorkletDirective=Ee;function Se(e,t){t.opts.processNestedWorklets&&e.traverse({[tr.WorkletizableFunction](n,i){Ee(n,i)}},t);let r=(0,nr.makeWorkletFactoryCall)(e,t);or(e,r)}C.processWorklet=Se;function ir(e){return e.some(t=>(0,ce.isDirectiveLiteral)(t.value)&&t.value.value==="worklet")}function or(e,t){var r;if(e.isObjectMethod())Ie(e,t);else{let n="id"in e.node?(r=e.node.id)===null||r===void 0?void 0:r.name:void 0;(0,rr.replaceWithFactoryCall)(e,n,t)}}function Ie(e,t){let r=(0,ce.objectProperty)(e.node.key,t);e.replaceWith(r)}C.substituteObjectMethodWithObjectProperty=Ie});var we=f(Y=>{"use strict";Object.defineProperty(Y,"__esModule",{value:!0});Y.isGestureHandlerEventCallback=void 0;var S=require("@babel/types"),sr=new Set(["Tap","Pan","Pinch","Rotation","Fling","LongPress","ForceTouch","Native","Manual","Race","Simultaneous","Exclusive","Hover"]),ar=new Set(["onBegin","onStart","onEnd","onFinalize","onUpdate","onChange","onTouchesDown","onTouchesMove","onTouchesUp","onTouchesCancelled"]);function lr(e){return(0,S.isCallExpression)(e.parent)&&(0,S.isExpression)(e.parent.callee)&&cr(e.parent.callee)}Y.isGestureHandlerEventCallback=lr;function cr(e){return(0,S.isMemberExpression)(e)&&(0,S.isIdentifier)(e.property)&&ar.has(e.property.name)&&xe(e.object)}function xe(e){return!!(ur(e)||(0,S.isCallExpression)(e)&&(0,S.isMemberExpression)(e.callee)&&xe(e.callee.object))}function ur(e){return(0,S.isCallExpression)(e)&&(0,S.isMemberExpression)(e.callee)&&(0,S.isIdentifier)(e.callee.object)&&e.callee.object.name==="Gesture"&&(0,S.isIdentifier)(e.callee.property)&&sr.has(e.callee.property.name)}});var Ce=f(K=>{"use strict";Object.defineProperty(K,"__esModule",{value:!0});K.isLayoutAnimationCallback=void 0;var I=require("@babel/types"),dr=new Set(["BounceIn","BounceInDown","BounceInLeft","BounceInRight","BounceInUp","BounceOut","BounceOutDown","BounceOutLeft","BounceOutRight","BounceOutUp","FadeIn","FadeInDown","FadeInLeft","FadeInRight","FadeInUp","FadeOut","FadeOutDown","FadeOutLeft","FadeOutRight","FadeOutUp","FlipInEasyX","FlipInEasyY","FlipInXDown","FlipInXUp","FlipInYLeft","FlipInYRight","FlipOutEasyX","FlipOutEasyY","FlipOutXDown","FlipOutXUp","FlipOutYLeft","FlipOutYRight","LightSpeedInLeft","LightSpeedInRight","LightSpeedOutLeft","LightSpeedOutRight","PinwheelIn","PinwheelOut","RollInLeft","RollInRight","RollOutLeft","RollOutRight","RotateInDownLeft","RotateInDownRight","RotateInUpLeft","RotateInUpRight","RotateOutDownLeft","RotateOutDownRight","RotateOutUpLeft","RotateOutUpRight","SlideInDown","SlideInLeft","SlideInRight","SlideInUp","SlideOutDown","SlideOutLeft","SlideOutRight","SlideOutUp","StretchInX","StretchInY","StretchOutX","StretchOutY","ZoomIn","ZoomInDown","ZoomInEasyDown","ZoomInEasyUp","ZoomInLeft","ZoomInRight","ZoomInRotate","ZoomInUp","ZoomOut","ZoomOutDown","ZoomOutEasyDown","ZoomOutEasyUp","ZoomOutLeft","ZoomOutRight","ZoomOutRotate","ZoomOutUp"]),fr=new Set(["Layout","LinearTransition","SequencedTransition","FadingTransition","JumpingTransition","CurvedTransition","EntryExitTransition"]),We=new Set([...dr,...fr]),pr=new Set(["build","duration","delay","getDuration","randomDelay","getDelay","getDelayFunction"]),br=new Set(["easing","rotate","springify","damping","mass","stiffness","overshootClamping","restDisplacementThreshold","restSpeedThreshold","withInitialValues","getAnimationAndConfig"]),mr=new Set(["easingX","easingY","easingWidth","easingHeight","entering","exiting","reverse"]),yr=new Set([...pr,...br,...mr]),kr=new Set(["withCallback"]);function gr(e){return(0,I.isCallExpression)(e.parent)&&(0,I.isExpression)(e.parent.callee)&&_r(e.parent.callee)}K.isLayoutAnimationCallback=gr;function _r(e){return(0,I.isMemberExpression)(e)&&(0,I.isIdentifier)(e.property)&&kr.has(e.property.name)&&Fe(e.object)}function Fe(e){return(0,I.isIdentifier)(e)&&We.has(e.name)?!0:!!((0,I.isNewExpression)(e)&&(0,I.isIdentifier)(e.callee)&&We.has(e.callee.name)||(0,I.isCallExpression)(e)&&(0,I.isMemberExpression)(e.callee)&&(0,I.isIdentifier)(e.callee.property)&&yr.has(e.callee.property.name)&&Fe(e.callee.object))}});var je=f(Q=>{"use strict";Object.defineProperty(Q,"__esModule",{value:!0});Q.findReferencedWorklet=void 0;var L=E();function Or(e,t,r){let n=e.node.name,s=e.scope.getBinding(n);return s?t&&s.path.isFunctionDeclaration()?s.path:s.constant?vr(s,t,r):hr(s,t,r):void 0}Q.findReferencedWorklet=Or;function vr(e,t,r){let n=e.path;if(!n.isVariableDeclarator())return;let i=n.get("init");if(t&&(0,L.isWorkletizableFunctionPath)(i)||r&&(0,L.isWorkletizableObjectPath)(i))return i}function hr(e,t,r){let n=e.constantViolations.reverse().find(s=>s.isAssignmentExpression()&&(t&&(0,L.isWorkletizableFunctionPath)(s.get("right"))||r&&(0,L.isWorkletizableObjectPath)(s.get("right"))));if(!n||!n.isAssignmentExpression())return;let i=n.get("right");if(t&&(0,L.isWorkletizableFunctionPath)(i)||r&&(0,L.isWorkletizableObjectPath)(i))return i}});var Re=f(ee=>{"use strict";Object.defineProperty(ee,"__esModule",{value:!0});ee.processWorkletizableObject=void 0;var Er=E(),De=$();function Sr(e,t){let r=e.get("properties");for(let n of r)if(n.isObjectMethod())(0,De.processWorklet)(n,t);else if(n.isObjectProperty()){let i=n.get("value");(0,Er.isWorkletizableFunctionPath)(i)&&(0,De.processWorklet)(i,t)}else throw new Error(`[Reanimated] '${n.type}' as to-be workletized argument is not supported for object hooks.`)}ee.processWorkletizableObject=Sr});var qe=f(z=>{"use strict";Object.defineProperty(z,"__esModule",{value:!0});z.processCalleesAutoworkletizableCallbacks=z.processIfAutoworkletizableCallback=void 0;var Ir=require("@babel/types"),te=E(),Me=$(),xr=we(),wr=Ce(),Wr=je(),Fr=Re(),Ae=new Set(["useAnimatedGestureHandler","useAnimatedScrollHandler"]),Pe=new Set(["useFrameCallback","useAnimatedStyle","useAnimatedProps","createAnimatedPropAdapter","useDerivedValue","useAnimatedScrollHandler","useAnimatedReaction","useWorkletCallback","withTiming","withSpring","withDecay","withRepeat","runOnUI","executeOnUIRuntimeSync"]),Cr=new Map([["useAnimatedGestureHandler",[0]],["useFrameCallback",[0]],["useAnimatedStyle",[0]],["useAnimatedProps",[0]],["createAnimatedPropAdapter",[0]],["useDerivedValue",[0]],["useAnimatedScrollHandler",[0]],["useAnimatedReaction",[0,1]],["useWorkletCallback",[0]],["withTiming",[2]],["withSpring",[2]],["withDecay",[1]],["withRepeat",[3]],["runOnUI",[0]],["executeOnUIRuntimeSync",[0]]]);function jr(e,t){return(0,xr.isGestureHandlerEventCallback)(e)||(0,wr.isLayoutAnimationCallback)(e)?((0,Me.processWorklet)(e,t),!0):!1}z.processIfAutoworkletizableCallback=jr;function Dr(e,t){let r=(0,Ir.isSequenceExpression)(e.node.callee)?e.node.callee.expressions[e.node.callee.expressions.length-1]:e.node.callee,n="name"in r?r.name:"property"in r&&"name"in r.property?r.property.name:void 0;if(n!==void 0&&(Pe.has(n)||Ae.has(n))){let i=Pe.has(n),s=Ae.has(n),l=Cr.get(n),u=e.get("arguments").filter((p,k)=>l.includes(k));Rr(u,t,i,s)}}z.processCalleesAutoworkletizableCallbacks=Dr;function Rr(e,t,r,n){e.forEach(i=>{let s=Ar(i,r,n);s&&((0,te.isWorkletizableFunctionPath)(s)?(0,Me.processWorklet)(s,t):(0,te.isWorkletizableObjectPath)(s)&&(0,Fr.processWorkletizableObject)(s,t))})}function Ar(e,t,r){if(t&&(0,te.isWorkletizableFunctionPath)(e)||r&&(0,te.isWorkletizableObjectPath)(e))return e;if(e.isReferencedIdentifier()&&e.isIdentifier())return(0,Wr.findReferencedWorklet)(e,t,r)}});var ue=f(h=>{"use strict";Object.defineProperty(h,"__esModule",{value:!0});h.isContextObject=h.processIfWorkletContextObject=h.contextObjectMarker=void 0;var v=require("@babel/types");h.contextObjectMarker="__workletContextObject";function Pr(e,t){return Le(e.node)?(qr(e.node),Mr(e.node),!0):!1}h.processIfWorkletContextObject=Pr;function Le(e){return e.properties.some(t=>(0,v.isObjectProperty)(t)&&(0,v.isIdentifier)(t.key)&&t.key.name===h.contextObjectMarker)}h.isContextObject=Le;function Mr(e){let t=(0,v.functionExpression)(null,[],(0,v.blockStatement)([(0,v.returnStatement)((0,v.cloneNode)(e))],[(0,v.directive)((0,v.directiveLiteral)("worklet"))]));e.properties.push((0,v.objectProperty)((0,v.identifier)(`${h.contextObjectMarker}Factory`),t))}function qr(e){e.properties=e.properties.filter(t=>!((0,v.isObjectProperty)(t)&&(0,v.isIdentifier)(t.key)&&t.key.name===h.contextObjectMarker))}});var Be=f(T=>{"use strict";Object.defineProperty(T,"__esModule",{value:!0});T.isImplicitContextObject=T.processIfWorkletFile=void 0;var b=require("@babel/types"),ze=E(),Te=ue();function Lr(e,t){return e.node.directives.some(r=>r.value.value==="worklet")?(e.node.directives=e.node.directives.filter(r=>r.value.value!=="worklet"),zr(e),!0):!1}T.processIfWorkletFile=Lr;function zr(e){let t=e.get("body");Vr(e.node),t.forEach(r=>{let n=Tr(r);de(n)})}function Tr(e){return e.isExportNamedDeclaration()||e.isExportDefaultDeclaration()?e.get("declaration"):e}function de(e){(0,ze.isWorkletizableFunctionPath)(e)?(e.isArrowFunctionExpression()&&Br(e.node),Ne(e.node.body)):(0,ze.isWorkletizableObjectPath)(e)?Ue(e)?Gr(e.node):Ur(e):e.isVariableDeclaration()?Nr(e):e.isClassDeclaration()&&Zr(e.node.body)}function Nr(e){e.get("declarations").forEach(r=>{let n=r.get("init");n.isExpression()&&de(n)})}function Ur(e){e.get("properties").forEach(r=>{if(r.isObjectMethod())Ne(r.node.body);else if(r.isObjectProperty()){let n=r.get("value");de(n)}})}function Br(e){(0,b.isBlockStatement)(e.body)||(e.body=(0,b.blockStatement)([(0,b.returnStatement)(e.body)]))}function Ne(e){e.directives.some(t=>t.value.value==="worklet")||e.directives.push((0,b.directive)((0,b.directiveLiteral)("worklet")))}function Gr(e){e.properties.some(t=>(0,b.isObjectProperty)(t)&&(0,b.isIdentifier)(t.key)&&t.key.name===Te.contextObjectMarker)||e.properties.push((0,b.objectProperty)((0,b.identifier)(`${Te.contextObjectMarker}`),(0,b.booleanLiteral)(!0)))}function Ue(e){return e.get("properties").some(r=>r.isObjectMethod()?Hr(r):!1)}T.isImplicitContextObject=Ue;function Hr(e){let t=!1;return e.traverse({ThisExpression(r){t=!0,r.stop()}}),t}function Zr(e){e.body.push((0,b.classProperty)((0,b.identifier)("__workletClass"),(0,b.booleanLiteral)(!0)))}function Vr(e){let t=e.body,r=t.length,n=0;for(;n<r;){let i=t[n];if(!Jr(i)){n++;continue}let s=t.splice(n,1);t.push(...s),r--}}function Jr(e){return(0,b.isExpressionStatement)(e)&&(0,b.isAssignmentExpression)(e.expression)&&(0,b.isMemberExpression)(e.expression.left)&&(0,b.isIdentifier)(e.expression.left.object)&&e.expression.left.object.name==="exports"}});var Ge=f(re=>{"use strict";Object.defineProperty(re,"__esModule",{value:!0});re.processInlineStylesWarning=void 0;var d=require("@babel/types"),Xr=D(),fe=require("assert");function $r(e){return(0,d.callExpression)((0,d.arrowFunctionExpression)([],(0,d.blockStatement)([(0,d.expressionStatement)((0,d.callExpression)((0,d.memberExpression)((0,d.identifier)("console"),(0,d.identifier)("warn")),[(0,d.callExpression)((0,d.memberExpression)((0,d.callExpression)((0,d.identifier)("require"),[(0,d.stringLiteral)("react-native-reanimated")]),(0,d.identifier)("getUseOfValueInStyleWarning")),[])])),(0,d.returnStatement)(e.node)])),[])}function Yr(e){e.isMemberExpression()&&(0,d.isIdentifier)(e.node.property)&&e.node.property.name==="value"&&e.replaceWith($r(e))}function Kr(e){if((0,d.isArrayExpression)(e.node)){let t=e.get("elements");(0,fe.strict)(Array.isArray(t),"[Reanimated] `elements` should be an array.");for(let r of t)r.isObjectExpression()&&pe(r)}}function pe(e){let t=e.get("properties");for(let r of t)if(r.isObjectProperty()){let n=r.get("value");(0,d.isIdentifier)(r.node.key)&&r.node.key.name==="transform"?Kr(n):Yr(n)}}function Qr(e,t){if((0,Xr.isRelease)()||t.opts.disableInlineStylesWarning||e.node.name.name!=="style"||!(0,d.isJSXExpressionContainer)(e.node.value))return;let r=e.get("value").get("expression");if((0,fe.strict)(!Array.isArray(r),"[Reanimated] `expression` should not be an array."),r.isArrayExpression()){let n=r.get("elements");(0,fe.strict)(Array.isArray(n),"[Reanimated] `elements` should be an array.");for(let i of n)i.isObjectExpression()&&pe(i)}else r.isObjectExpression()&&pe(r)}re.processInlineStylesWarning=Qr});var Ze=f(ne=>{"use strict";Object.defineProperty(ne,"__esModule",{value:!0});ne.substituteWebCallExpression=void 0;var He=require("@babel/types");function en(e){let t=e.node.callee;if((0,He.isIdentifier)(t)){let r=t.name;(r==="isWeb"||r==="shouldBeUseWeb")&&e.replaceWith((0,He.booleanLiteral)(!0))}}ne.substituteWebCallExpression=en});var $e=f(N=>{"use strict";var Ve=N&&N.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(N,"__esModule",{value:!0});N.processIfWorkletClass=void 0;var tn=Ve(require("@babel/generator")),rn=Ve(require("@babel/traverse")),a=require("@babel/types"),ie=require("assert"),nn=E(),on=D(),sn=J(),Je="__workletClass";function an(e,t){return!e.node.id||!pn(e.node.body)?!1:(bn(e.node.body),ln(e,t),!0)}N.processIfWorkletClass=an;function ln(e,t){(0,ie.strict)(e.node.id);let r=e.node.id.name,n=cn(e.node,t);mn(n),un(n.program.body),dn(n.program.body,r),n.program.body.push((0,a.returnStatement)((0,a.identifier)(r)));let i=(0,a.functionExpression)(null,[],(0,a.blockStatement)([...n.program.body])),s=(0,a.callExpression)(i,[]);(0,on.replaceWithFactoryCall)(e,r,s)}function cn(e,t){let r=(0,tn.default)(e).code,n=(0,sn.workletTransformSync)(r,{extraPlugins:["@babel/plugin-transform-class-properties","@babel/plugin-transform-classes","@babel/plugin-transform-unicode-regex"],filename:t.file.opts.filename,ast:!0,babelrc:!1,configFile:!1});return(0,ie.strict)(n&&n.ast),n.ast}function un(e){e.forEach(t=>{if((0,a.isFunctionDeclaration)(t)){let r=(0,a.directive)((0,a.directiveLiteral)("worklet"));t.body.directives.push(r)}})}function dn(e,t){let r=t+nn.workletClassFactorySuffix,n=fn(e,t),s=e[n].declarations[0].init,l=(0,a.functionDeclaration)((0,a.identifier)(r),[],(0,a.blockStatement)([(0,a.variableDeclaration)("const",[(0,a.variableDeclarator)((0,a.identifier)(t),s)]),(0,a.expressionStatement)((0,a.assignmentExpression)("=",(0,a.memberExpression)((0,a.identifier)(t),(0,a.identifier)(r)),(0,a.identifier)(r))),(0,a.returnStatement)((0,a.identifier)(t))],[(0,a.directive)((0,a.directiveLiteral)("worklet"))])),u=(0,a.variableDeclaration)("const",[(0,a.variableDeclarator)((0,a.identifier)(t),(0,a.callExpression)((0,a.identifier)(r),[]))]);e.splice(n,1,l,u)}function fn(e,t){let r=e.findIndex(n=>(0,a.isVariableDeclaration)(n)&&n.declarations.some(i=>(0,a.isIdentifier)(i.id)&&i.id.name===t));return(0,ie.strict)(r>=0),r}function pn(e){return e.body.some(t=>(0,a.isClassProperty)(t)&&(0,a.isIdentifier)(t.key)&&t.key.name===Je)}function bn(e){e.body=e.body.filter(t=>!(0,a.isClassProperty)(t)||!(0,a.isIdentifier)(t.key)||t.key.name!==Je)}function mn(e){let t=yn(e),r=kn(t),n=t.map(u=>u.index),i=r.map(u=>u.index),s=e.program.body,l=[...s];for(let u=0;u<t.length;u++){let p=i[u],k=n[u],_=l[p];s[k]=_}}function yn(e){let t=[];return(0,rn.default)(e,{Program:{enter:r=>{r.get("body").forEach((i,s)=>{var l;let u=i.getBindingIdentifiers();if(!i.isFunctionDeclaration()||!(!((l=i.node.id)===null||l===void 0)&&l.name))return;let p={name:i.node.id.name,index:s,dependencies:new Set};t.push(p),i.traverse({Identifier(k){gn(k,u,i)&&p.dependencies.add(k.node.name)}})})}}}),t}function kn(e){let t=[],r=new Set;for(let n of e)Xe(n,e,t,r);return t}function Xe(e,t,r,n){if(n.has(e.name))throw new Error("Cycle detected. This should never happen.");if(!r.find(i=>i.name===e.name)){n.add(e.name);for(let i of e.dependencies)if(!r.find(s=>s.name===i)){let s=t.find(l=>l.name===i);(0,ie.strict)(s),Xe(s,t,r,n)}r.push(e),n.delete(e.name)}}function gn(e,t,r){return e.isReferencedIdentifier()&&!(e.node.name in t)&&!r.scope.hasOwnBinding(e.node.name)&&r.scope.hasReference(e.node.name)}});Object.defineProperty(exports,"__esModule",{value:!0});var Ye=qe(),_n=ue(),On=Be(),vn=V(),hn=Ge(),En=E(),Sn=D(),In=Ze(),xn=$(),wn=$e();module.exports=function(){function e(t){try{t()}catch(r){throw new Error(`[Reanimated] Babel plugin exception: ${r}`)}}return{pre(t){e(()=>{t.workletNumber=1,(0,vn.initializeGlobals)(),Sn.addCustomGlobals.call(this)})},visitor:{CallExpression:{enter(t,r){e(()=>{(0,Ye.processCalleesAutoworkletizableCallbacks)(t,r),r.opts.substituteWebPlatformChecks&&(0,In.substituteWebCallExpression)(t)})}},[En.WorkletizableFunction]:{enter(t,r){e(()=>{(0,xn.processIfWithWorkletDirective)(t,r)||(0,Ye.processIfAutoworkletizableCallback)(t,r)})}},ObjectExpression:{enter(t,r){e(()=>{(0,_n.processIfWorkletContextObject)(t,r)})}},ClassDeclaration:{enter(t,r){e(()=>{(0,wn.processIfWorkletClass)(t,r)})}},Program:{enter(t,r){e(()=>{r.workletNumber=1,(0,On.processIfWorkletFile)(t,r)})}},JSXAttribute:{enter(t,r){e(()=>(0,hn.processInlineStylesWarning)(t,r))}}}}};
+"use strict";var Ke=Object.create;var me=Object.defineProperty;var Qe=Object.getOwnPropertyDescriptor;var et=Object.getOwnPropertyNames;var tt=Object.getPrototypeOf,rt=Object.prototype.hasOwnProperty;var f=(e,t)=>()=>(t||e((t={exports:{}}).exports,t),t.exports);var nt=(e,t,r,n)=>{if(t&&typeof t=="object"||typeof t=="function")for(let i of et(t))!rt.call(e,i)&&i!==r&&me(e,i,{get:()=>t[i],enumerable:!(n=Qe(t,i))||n.enumerable});return e};var P=(e,t,r)=>(r=e!=null?Ke(tt(e)):{},nt(t||!e||!e.__esModule?me(r,"default",{value:e,enumerable:!0}):r,e));var E=f(m=>{"use strict";Object.defineProperty(m,"__esModule",{value:!0});m.workletClassFactorySuffix=m.isWorkletizableObjectNode=m.isWorkletizableObjectPath=m.isWorkletizableFunctionNode=m.isWorkletizableFunctionPath=m.WorkletizableObject=m.WorkletizableFunction=void 0;var B=require("@babel/types");m.WorkletizableFunction="FunctionDeclaration|FunctionExpression|ArrowFunctionExpression|ObjectMethod";m.WorkletizableObject="ObjectExpression";function it(e){return e.isFunctionDeclaration()||e.isFunctionExpression()||e.isArrowFunctionExpression()||e.isObjectMethod()}m.isWorkletizableFunctionPath=it;function ot(e){return(0,B.isFunctionDeclaration)(e)||(0,B.isFunctionExpression)(e)||(0,B.isArrowFunctionExpression)(e)||(0,B.isObjectMethod)(e)}m.isWorkletizableFunctionNode=ot;function st(e){return e.isObjectExpression()}m.isWorkletizableObjectPath=st;function at(e){return(0,B.isObjectExpression)(e)}m.isWorkletizableObjectNode=at;m.workletClassFactorySuffix="__classFactory"});var V=f(w=>{"use strict";Object.defineProperty(w,"__esModule",{value:!0});w.initializeGlobals=w.globals=w.defaultGlobals=void 0;var lt=["globalThis","Infinity","NaN","undefined","eval","isFinite","isNaN","parseFloat","parseInt","decodeURI","decodeURIComponent","encodeURI","encodeURIComponent","escape","unescape","Object","Function","Boolean","Symbol","Error","AggregateError","EvalError","RangeError","ReferenceError","SyntaxError","TypeError","URIError","InternalError","Number","BigInt","Math","Date","String","RegExp","Array","Int8Array","Uint8Array","Uint8ClampedArray","Int16Array","Uint16Array","Int32Array","Uint32Array","BigInt64Array","BigUint64Array","Float32Array","Float64Array","Map","Set","WeakMap","WeakSet","ArrayBuffer","SharedArrayBuffer","DataView","Atomics","JSON","WeakRef","FinalizationRegistry","Iterator","AsyncIterator","Promise","GeneratorFunction","AsyncGeneratorFunction","Generator","AsyncGenerator","AsyncFunction","Reflect","Proxy","Intl","null","this","global","window","globalThis","console","performance","queueMicrotask","requestAnimationFrame","setImmediate","arguments","HermesInternal","_WORKLET","ReanimatedError","__reanimatedLoggerConfig"],ct=["_IS_FABRIC","_log","_toString","_scheduleHostFunctionOnJS","_scheduleRemoteFunctionOnJS","_scheduleOnRuntime","_makeShareableClone","_updatePropsPaper","_updatePropsFabric","_removeFromPropsRegistry","_measurePaper","_measureFabric","_scrollToPaper","_dispatchCommandPaper","_dispatchCommandFabric","_setGestureState","_notifyAboutProgress","_notifyAboutEnd","_runOnUIQueue","_getAnimationTimestamp"];w.defaultGlobals=new Set(lt.concat(ct));function ut(){w.globals=new Set(w.defaultGlobals)}w.initializeGlobals=ut});var D=f(W=>{"use strict";Object.defineProperty(W,"__esModule",{value:!0});W.replaceWithFactoryCall=W.addCustomGlobals=W.isRelease=void 0;var G=require("@babel/types"),dt=V();function ft(){var e,t;let r=/(prod|release|stag[ei])/i;return!!(!((e=process.env.BABEL_ENV)===null||e===void 0)&&e.match(r)||!((t=process.env.NODE_ENV)===null||t===void 0)&&t.match(r))}W.isRelease=ft;function pt(){this.opts&&Array.isArray(this.opts.globals)&&this.opts.globals.forEach(e=>{dt.globals.add(e)})}W.addCustomGlobals=pt;function bt(e,t,r){if(!t||!mt(e))e.replaceWith(r);else{let n=(0,G.variableDeclaration)("const",[(0,G.variableDeclarator)((0,G.identifier)(t),r)]);e.replaceWith(n)}}W.replaceWithFactoryCall=bt;function mt(e){return(0,G.isScopable)(e.parent)||(0,G.isExportNamedDeclaration)(e.parent)}});var J=f(M=>{"use strict";var yt=M&&M.__rest||function(e,t){var r={};for(var n in e)Object.prototype.hasOwnProperty.call(e,n)&&t.indexOf(n)<0&&(r[n]=e[n]);if(e!=null&&typeof Object.getOwnPropertySymbols=="function")for(var i=0,n=Object.getOwnPropertySymbols(e);i<n.length;i++)t.indexOf(n[i])<0&&Object.prototype.propertyIsEnumerable.call(e,n[i])&&(r[n[i]]=e[n[i]]);return r};Object.defineProperty(M,"__esModule",{value:!0});M.workletTransformSync=void 0;var kt=require("@babel/core");function gt(e,t){let{extraPlugins:r=[],extraPresets:n=[]}=t,i=yt(t,["extraPlugins","extraPresets"]);return(0,kt.transformSync)(e,Object.assign(Object.assign({},i),{plugins:[...Ot,...r],presets:[..._t,...n]}))}M.workletTransformSync=gt;var _t=[require.resolve("@babel/preset-typescript")],Ot=[]});var ge=f(O=>{"use strict";var vt=O&&O.__createBinding||(Object.create?function(e,t,r,n){n===void 0&&(n=r);var i=Object.getOwnPropertyDescriptor(t,r);(!i||("get"in i?!t.__esModule:i.writable||i.configurable))&&(i={enumerable:!0,get:function(){return t[r]}}),Object.defineProperty(e,n,i)}:function(e,t,r,n){n===void 0&&(n=r),e[n]=t[r]}),ht=O&&O.__setModuleDefault||(Object.create?function(e,t){Object.defineProperty(e,"default",{enumerable:!0,value:t})}:function(e,t){e.default=t}),ye=O&&O.__importStar||function(e){if(e&&e.__esModule)return e;var t={};if(e!=null)for(var r in e)r!=="default"&&Object.prototype.hasOwnProperty.call(e,r)&&vt(t,e,r);return ht(t,e),t},Et=O&&O.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(O,"__esModule",{value:!0});O.buildWorkletString=void 0;var ke=require("@babel/core"),St=Et(require("@babel/generator")),c=require("@babel/types"),H=require("assert"),It=ye(require("convert-source-map")),xt=ye(require("fs")),wt=E(),Wt=D(),Ft=J(),Ct="mock source map";function jt(e,t,r,n,i){Dt(e,n);let s=e.program.body.find(g=>(0,c.isFunctionDeclaration)(g))||e.program.body.find(g=>(0,c.isExpressionStatement)(g))||void 0;(0,H.strict)(s,"[Reanimated] `draftExpression` is undefined.");let l=(0,c.isFunctionDeclaration)(s)?s:s.expression;(0,H.strict)("params"in l,"'params' property is undefined in 'expression'"),(0,H.strict)((0,c.isBlockStatement)(l.body),"[Reanimated] `expression.body` is not a `BlockStatement`");let u=new Set;(0,ke.traverse)(e,{NewExpression(g){if(!(0,c.isIdentifier)(g.node.callee))return;let x=g.node.callee.name;if(!r.some(Z=>Z.name===x)||u.has(x))return;let A=r.findIndex(Z=>Z.name===x);r.splice(A,1);let oe=x+wt.workletClassFactorySuffix;r.push((0,c.identifier)(oe)),(0,c.assertBlockStatement)(l.body),l.body.body.unshift((0,c.variableDeclaration)("const",[(0,c.variableDeclarator)((0,c.identifier)(x),(0,c.callExpression)((0,c.identifier)(oe),[]))])),u.add(x)}});let p=(0,c.functionExpression)((0,c.identifier)(n),l.params,l.body,l.generator,l.async),k=(0,St.default)(p).code;(0,H.strict)(i,"[Reanimated] `inputMap` is undefined.");let _=!((0,Wt.isRelease)()||t.opts.disableSourceMaps);if(_){i.sourcesContent=[];for(let g of i.sources)i.sourcesContent.push(xt.readFileSync(g).toString("utf-8"))}let R=(0,Ft.workletTransformSync)(k,{filename:t.file.opts.filename,extraPlugins:[Mt(r)],compact:!0,sourceMaps:_,inputSourceMap:i,ast:!1,babelrc:!1,configFile:!1,comments:!1});(0,H.strict)(R,"[Reanimated] `transformed` is null.");let j;return _&&(Rt()?j=Ct:(j=It.fromObject(R.map).toObject(),delete j.sourcesContent)),[R.code,JSON.stringify(j)]}O.buildWorkletString=jt;function Dt(e,t){(0,ke.traverse)(e,{FunctionExpression(r){if(!r.node.id){r.stop();return}let n=r.node.id.name;r.scope.rename(n,t)}})}function Rt(){return process.env.REANIMATED_JEST_SHOULD_MOCK_SOURCE_MAP==="1"}function At(e,t,r){t.length===0||!(0,c.isProgram)(e.parent)||(0,c.isExpression)(e.node.body)||e.node.body.body.unshift(r)}function Pt(e){var t;(0,c.isProgram)(e.parent)&&!(0,c.isArrowFunctionExpression)(e.node)&&!(0,c.isObjectMethod)(e.node)&&e.node.id&&e.scope.parent&&((t=e.scope.parent.bindings[e.node.id.name])===null||t===void 0?void 0:t.references)>0&&e.node.body.body.unshift((0,c.variableDeclaration)("const",[(0,c.variableDeclarator)((0,c.identifier)(e.node.id.name),(0,c.memberExpression)((0,c.thisExpression)(),(0,c.identifier)("_recur")))]))}function Mt(e){let t=(0,c.variableDeclaration)("const",[(0,c.variableDeclarator)((0,c.objectPattern)(e.map(r=>(0,c.objectProperty)((0,c.identifier)(r.name),(0,c.identifier)(r.name),!1,!0))),(0,c.memberExpression)((0,c.thisExpression)(),(0,c.identifier)("__closure")))]);return{visitor:{"FunctionDeclaration|FunctionExpression|ArrowFunctionExpression|ObjectMethod":r=>{At(r,e,t),Pt(r)}}}}});var ve=f(q=>{"use strict";var qt=q&&q.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(q,"__esModule",{value:!0});q.makeWorkletFactory=void 0;var Lt=require("@babel/core"),zt=qt(require("@babel/generator")),o=require("@babel/types"),F=require("assert"),Oe=require("path"),Tt=V(),_e=E(),le=D(),Nt=ge(),Ut=J(),Bt=require("../package.json").version,Gt="x.y.z";function Ht(e,t){Zt(e),(0,F.strict)(t.file.opts.filename,"[Reanimated] `state.file.opts.filename` is undefined.");let r=(0,zt.default)(e.node,{sourceMaps:!0,sourceFileName:t.file.opts.filename});r.code="("+(e.isObjectMethod()?"function ":"")+r.code+`
+)`;let n=(0,Ut.workletTransformSync)(r.code,{extraPlugins:Yt,filename:t.file.opts.filename,ast:!0,babelrc:!1,configFile:!1,inputSourceMap:r.map});(0,F.strict)(n,"[Reanimated] `transformed` is undefined."),(0,F.strict)(n.ast,"[Reanimated] `transformed.ast` is undefined.");let i=$t(n.ast,e),s=(0,o.cloneNode)(e.node),l=(0,o.isBlockStatement)(s.body)?(0,o.functionExpression)(null,s.params,s.body,s.generator,s.async):s,{workletName:u,reactName:p}=Xt(e,t),[k,_]=(0,Nt.buildWorkletString)(n.ast,t,i,u,n.map);(0,F.strict)(k,"[Reanimated] `funString` is undefined.");let R=Jt(k),j=1;i.length>0&&(j-=i.length+2);let g=e.parentPath.isProgram()?e:e.findParent(y=>{var se,ae;return(ae=(se=y.parentPath)===null||se===void 0?void 0:se.isProgram())!==null&&ae!==void 0?ae:!1});(0,F.strict)(g,"[Reanimated] `pathForStringDefinitions` is null."),(0,F.strict)(g.parentPath,"[Reanimated] `pathForStringDefinitions.parentPath` is null.");let x=g.parentPath.scope.generateUidIdentifier(`worklet_${R}_init_data`),A=(0,o.objectExpression)([(0,o.objectProperty)((0,o.identifier)("code"),(0,o.stringLiteral)(k))]);if(!(0,le.isRelease)()){let y=t.file.opts.filename;t.opts.relativeSourceLocation&&(y=(0,Oe.relative)(t.cwd,y),_=_==null?void 0:_.replace(t.file.opts.filename,y)),A.properties.push((0,o.objectProperty)((0,o.identifier)("location"),(0,o.stringLiteral)(y)))}_&&A.properties.push((0,o.objectProperty)((0,o.identifier)("sourceMap"),(0,o.stringLiteral)(_))),!(0,le.isRelease)()&&A.properties.push((0,o.objectProperty)((0,o.identifier)("version"),(0,o.stringLiteral)(Vt()?Gt:Bt)));let be=!t.opts.omitNativeOnlyData;be&&g.insertBefore((0,o.variableDeclaration)("const",[(0,o.variableDeclarator)(x,A)])),(0,F.strict)(!(0,o.isFunctionDeclaration)(l),"[Reanimated] `funExpression` is a `FunctionDeclaration`."),(0,F.strict)(!(0,o.isObjectMethod)(l),"[Reanimated] `funExpression` is an `ObjectMethod`.");let U=[(0,o.variableDeclaration)("const",[(0,o.variableDeclarator)((0,o.identifier)(p),l)]),(0,o.expressionStatement)((0,o.assignmentExpression)("=",(0,o.memberExpression)((0,o.identifier)(p),(0,o.identifier)("__closure"),!1),(0,o.objectExpression)(i.map(y=>y.name.endsWith(_e.workletClassFactorySuffix)?(0,o.objectProperty)((0,o.identifier)(y.name),(0,o.memberExpression)((0,o.identifier)(y.name.slice(0,y.name.length-_e.workletClassFactorySuffix.length)),(0,o.identifier)(y.name))):(0,o.objectProperty)((0,o.identifier)(y.name),y,!1,!0))))),(0,o.expressionStatement)((0,o.assignmentExpression)("=",(0,o.memberExpression)((0,o.identifier)(p),(0,o.identifier)("__workletHash"),!1),(0,o.numericLiteral)(R)))];return be&&U.push((0,o.expressionStatement)((0,o.assignmentExpression)("=",(0,o.memberExpression)((0,o.identifier)(p),(0,o.identifier)("__initData"),!1),x))),(0,le.isRelease)()||(U.unshift((0,o.variableDeclaration)("const",[(0,o.variableDeclarator)((0,o.identifier)("_e"),(0,o.arrayExpression)([(0,o.newExpression)((0,o.memberExpression)((0,o.identifier)("global"),(0,o.identifier)("Error")),[]),(0,o.numericLiteral)(j),(0,o.numericLiteral)(-27)]))])),U.push((0,o.expressionStatement)((0,o.assignmentExpression)("=",(0,o.memberExpression)((0,o.identifier)(p),(0,o.identifier)("__stackDetails"),!1),(0,o.identifier)("_e"))))),U.push((0,o.returnStatement)((0,o.identifier)(p))),(0,o.functionExpression)(void 0,[],(0,o.blockStatement)(U))}q.makeWorkletFactory=Ht;function Zt(e){e.traverse({DirectiveLiteral(t){t.node.value==="worklet"&&t.getFunctionParent()===e&&t.parentPath.remove()}})}function Vt(){return process.env.REANIMATED_JEST_SHOULD_MOCK_VERSION==="1"}function Jt(e){let t=e.length,r=5381,n=52711;for(;t--;){let i=e.charCodeAt(t);r=r*33^i,n=n*33^i}return(r>>>0)*4096+(n>>>0)}function Xt(e,t){let r="unknownFile";if(t.file.opts.filename){let l=t.file.opts.filename;r=(0,Oe.basename)(l);let u=l.split("/"),p=u.indexOf("node_modules");p!==-1&&(r=`${u[p+1]}_${r}`)}let n=`${r}${t.workletNumber++}`,i="";(0,o.isObjectMethod)(e.node)&&(0,o.isIdentifier)(e.node.key)?i=e.node.key.name:((0,o.isFunctionDeclaration)(e.node)||(0,o.isFunctionExpression)(e.node))&&(0,o.isIdentifier)(e.node.id)&&(i=e.node.id.name);let s=i?(0,o.toIdentifier)(`${i}_${n}`):(0,o.toIdentifier)(n);return i=i||(0,o.toIdentifier)(n),{workletName:s,reactName:i}}function $t(e,t){let r=new Map,n=new Map;return(0,Lt.traverse)(e,{Identifier(i){if(!i.isReferencedIdentifier())return;let s=i.node.name;if(Tt.globals.has(s)||"id"in t.node&&t.node.id&&t.node.id.name===s)return;let l=i.parent;if((0,o.isMemberExpression)(l)&&l.property===i.node&&!l.computed||(0,o.isObjectProperty)(l)&&(0,o.isObjectExpression)(i.parentPath.parent)&&i.node!==l.value)return;let u=i.scope;for(;u!=null;){if(u.bindings[s]!=null)return;u=u.parent}r.set(s,i.node),n.set(s,!1)}}),t.traverse({Identifier(i){if(!i.isReferencedIdentifier())return;let s=r.get(i.node.name);!s||n.get(i.node.name)||(s.loc=i.node.loc,n.set(i.node.name,!0))}}),Array.from(r.values())}var Yt=[require.resolve("@babel/plugin-transform-shorthand-properties"),require.resolve("@babel/plugin-transform-arrow-functions"),require.resolve("@babel/plugin-transform-optional-chaining"),require.resolve("@babel/plugin-transform-nullish-coalescing-operator"),[require.resolve("@babel/plugin-transform-template-literals"),{loose:!0}]]});var he=f(X=>{"use strict";Object.defineProperty(X,"__esModule",{value:!0});X.makeWorkletFactoryCall=void 0;var Kt=require("@babel/types"),Qt=ve();function er(e,t){let r=(0,Qt.makeWorkletFactory)(e,t),n=(0,Kt.callExpression)(r,[]);return tr(e,n),n}X.makeWorkletFactoryCall=er;function tr(e,t){let r=e.node.loc;r&&(t.callee.loc={filename:r.filename,identifierName:r.identifierName,start:r.start,end:r.start})}});var $=f(C=>{"use strict";Object.defineProperty(C,"__esModule",{value:!0});C.substituteObjectMethodWithObjectProperty=C.processWorklet=C.processIfWithWorkletDirective=void 0;var ce=require("@babel/types"),rr=E(),nr=D(),ir=he();function Ee(e,t){return!(0,ce.isBlockStatement)(e.node.body)||!or(e.node.body.directives)?!1:(Se(e,t),!0)}C.processIfWithWorkletDirective=Ee;function Se(e,t){t.opts.processNestedWorklets&&e.traverse({[rr.WorkletizableFunction](n,i){Ee(n,i)}},t);let r=(0,ir.makeWorkletFactoryCall)(e,t);sr(e,r)}C.processWorklet=Se;function or(e){return e.some(t=>(0,ce.isDirectiveLiteral)(t.value)&&t.value.value==="worklet")}function sr(e,t){var r;if(e.isObjectMethod())Ie(e,t);else{let n="id"in e.node?(r=e.node.id)===null||r===void 0?void 0:r.name:void 0;(0,nr.replaceWithFactoryCall)(e,n,t)}}function Ie(e,t){let r=(0,ce.objectProperty)(e.node.key,t);e.replaceWith(r)}C.substituteObjectMethodWithObjectProperty=Ie});var we=f(Y=>{"use strict";Object.defineProperty(Y,"__esModule",{value:!0});Y.isGestureHandlerEventCallback=void 0;var S=require("@babel/types"),ar=new Set(["Tap","Pan","Pinch","Rotation","Fling","LongPress","ForceTouch","Native","Manual","Race","Simultaneous","Exclusive","Hover"]),lr=new Set(["onBegin","onStart","onEnd","onFinalize","onUpdate","onChange","onTouchesDown","onTouchesMove","onTouchesUp","onTouchesCancelled"]);function cr(e){return(0,S.isCallExpression)(e.parent)&&(0,S.isExpression)(e.parent.callee)&&ur(e.parent.callee)}Y.isGestureHandlerEventCallback=cr;function ur(e){return(0,S.isMemberExpression)(e)&&(0,S.isIdentifier)(e.property)&&lr.has(e.property.name)&&xe(e.object)}function xe(e){return!!(dr(e)||(0,S.isCallExpression)(e)&&(0,S.isMemberExpression)(e.callee)&&xe(e.callee.object))}function dr(e){return(0,S.isCallExpression)(e)&&(0,S.isMemberExpression)(e.callee)&&(0,S.isIdentifier)(e.callee.object)&&e.callee.object.name==="Gesture"&&(0,S.isIdentifier)(e.callee.property)&&ar.has(e.callee.property.name)}});var Ce=f(K=>{"use strict";Object.defineProperty(K,"__esModule",{value:!0});K.isLayoutAnimationCallback=void 0;var I=require("@babel/types"),fr=new Set(["BounceIn","BounceInDown","BounceInLeft","BounceInRight","BounceInUp","BounceOut","BounceOutDown","BounceOutLeft","BounceOutRight","BounceOutUp","FadeIn","FadeInDown","FadeInLeft","FadeInRight","FadeInUp","FadeOut","FadeOutDown","FadeOutLeft","FadeOutRight","FadeOutUp","FlipInEasyX","FlipInEasyY","FlipInXDown","FlipInXUp","FlipInYLeft","FlipInYRight","FlipOutEasyX","FlipOutEasyY","FlipOutXDown","FlipOutXUp","FlipOutYLeft","FlipOutYRight","LightSpeedInLeft","LightSpeedInRight","LightSpeedOutLeft","LightSpeedOutRight","PinwheelIn","PinwheelOut","RollInLeft","RollInRight","RollOutLeft","RollOutRight","RotateInDownLeft","RotateInDownRight","RotateInUpLeft","RotateInUpRight","RotateOutDownLeft","RotateOutDownRight","RotateOutUpLeft","RotateOutUpRight","SlideInDown","SlideInLeft","SlideInRight","SlideInUp","SlideOutDown","SlideOutLeft","SlideOutRight","SlideOutUp","StretchInX","StretchInY","StretchOutX","StretchOutY","ZoomIn","ZoomInDown","ZoomInEasyDown","ZoomInEasyUp","ZoomInLeft","ZoomInRight","ZoomInRotate","ZoomInUp","ZoomOut","ZoomOutDown","ZoomOutEasyDown","ZoomOutEasyUp","ZoomOutLeft","ZoomOutRight","ZoomOutRotate","ZoomOutUp"]),pr=new Set(["Layout","LinearTransition","SequencedTransition","FadingTransition","JumpingTransition","CurvedTransition","EntryExitTransition"]),We=new Set([...fr,...pr]),br=new Set(["build","duration","delay","getDuration","randomDelay","getDelay","getDelayFunction"]),mr=new Set(["easing","rotate","springify","damping","mass","stiffness","overshootClamping","restDisplacementThreshold","restSpeedThreshold","withInitialValues","getAnimationAndConfig"]),yr=new Set(["easingX","easingY","easingWidth","easingHeight","entering","exiting","reverse"]),kr=new Set([...br,...mr,...yr]),gr=new Set(["withCallback"]);function _r(e){return(0,I.isCallExpression)(e.parent)&&(0,I.isExpression)(e.parent.callee)&&Or(e.parent.callee)}K.isLayoutAnimationCallback=_r;function Or(e){return(0,I.isMemberExpression)(e)&&(0,I.isIdentifier)(e.property)&&gr.has(e.property.name)&&Fe(e.object)}function Fe(e){return(0,I.isIdentifier)(e)&&We.has(e.name)?!0:!!((0,I.isNewExpression)(e)&&(0,I.isIdentifier)(e.callee)&&We.has(e.callee.name)||(0,I.isCallExpression)(e)&&(0,I.isMemberExpression)(e.callee)&&(0,I.isIdentifier)(e.callee.property)&&kr.has(e.callee.property.name)&&Fe(e.callee.object))}});var je=f(Q=>{"use strict";Object.defineProperty(Q,"__esModule",{value:!0});Q.findReferencedWorklet=void 0;var L=E();function vr(e,t,r){let n=e.node.name,s=e.scope.getBinding(n);return s?t&&s.path.isFunctionDeclaration()?s.path:s.constant?hr(s,t,r):Er(s,t,r):void 0}Q.findReferencedWorklet=vr;function hr(e,t,r){let n=e.path;if(!n.isVariableDeclarator())return;let i=n.get("init");if(t&&(0,L.isWorkletizableFunctionPath)(i)||r&&(0,L.isWorkletizableObjectPath)(i))return i}function Er(e,t,r){let n=e.constantViolations.reverse().find(s=>s.isAssignmentExpression()&&(t&&(0,L.isWorkletizableFunctionPath)(s.get("right"))||r&&(0,L.isWorkletizableObjectPath)(s.get("right"))));if(!n||!n.isAssignmentExpression())return;let i=n.get("right");if(t&&(0,L.isWorkletizableFunctionPath)(i)||r&&(0,L.isWorkletizableObjectPath)(i))return i}});var Re=f(ee=>{"use strict";Object.defineProperty(ee,"__esModule",{value:!0});ee.processWorkletizableObject=void 0;var Sr=E(),De=$();function Ir(e,t){let r=e.get("properties");for(let n of r)if(n.isObjectMethod())(0,De.processWorklet)(n,t);else if(n.isObjectProperty()){let i=n.get("value");(0,Sr.isWorkletizableFunctionPath)(i)&&(0,De.processWorklet)(i,t)}else throw new Error(`[Reanimated] '${n.type}' as to-be workletized argument is not supported for object hooks.`)}ee.processWorkletizableObject=Ir});var qe=f(z=>{"use strict";Object.defineProperty(z,"__esModule",{value:!0});z.processCalleesAutoworkletizableCallbacks=z.processIfAutoworkletizableCallback=void 0;var xr=require("@babel/types"),te=E(),Me=$(),wr=we(),Wr=Ce(),Fr=je(),Cr=Re(),Ae=new Set(["useAnimatedGestureHandler","useAnimatedScrollHandler"]),Pe=new Set(["useFrameCallback","useAnimatedStyle","useAnimatedProps","createAnimatedPropAdapter","useDerivedValue","useAnimatedScrollHandler","useAnimatedReaction","useWorkletCallback","withTiming","withSpring","withDecay","withRepeat","runOnUI","executeOnUIRuntimeSync"]),jr=new Map([["useAnimatedGestureHandler",[0]],["useFrameCallback",[0]],["useAnimatedStyle",[0]],["useAnimatedProps",[0]],["createAnimatedPropAdapter",[0]],["useDerivedValue",[0]],["useAnimatedScrollHandler",[0]],["useAnimatedReaction",[0,1]],["useWorkletCallback",[0]],["withTiming",[2]],["withSpring",[2]],["withDecay",[1]],["withRepeat",[3]],["runOnUI",[0]],["executeOnUIRuntimeSync",[0]]]);function Dr(e,t){return(0,wr.isGestureHandlerEventCallback)(e)||(0,Wr.isLayoutAnimationCallback)(e)?((0,Me.processWorklet)(e,t),!0):!1}z.processIfAutoworkletizableCallback=Dr;function Rr(e,t){let r=(0,xr.isSequenceExpression)(e.node.callee)?e.node.callee.expressions[e.node.callee.expressions.length-1]:e.node.callee,n="name"in r?r.name:"property"in r&&"name"in r.property?r.property.name:void 0;if(n!==void 0&&(Pe.has(n)||Ae.has(n))){let i=Pe.has(n),s=Ae.has(n),l=jr.get(n),u=e.get("arguments").filter((p,k)=>l.includes(k));Ar(u,t,i,s)}}z.processCalleesAutoworkletizableCallbacks=Rr;function Ar(e,t,r,n){e.forEach(i=>{let s=Pr(i,r,n);s&&((0,te.isWorkletizableFunctionPath)(s)?(0,Me.processWorklet)(s,t):(0,te.isWorkletizableObjectPath)(s)&&(0,Cr.processWorkletizableObject)(s,t))})}function Pr(e,t,r){if(t&&(0,te.isWorkletizableFunctionPath)(e)||r&&(0,te.isWorkletizableObjectPath)(e))return e;if(e.isReferencedIdentifier()&&e.isIdentifier())return(0,Fr.findReferencedWorklet)(e,t,r)}});var ue=f(h=>{"use strict";Object.defineProperty(h,"__esModule",{value:!0});h.isContextObject=h.processIfWorkletContextObject=h.contextObjectMarker=void 0;var v=require("@babel/types");h.contextObjectMarker="__workletContextObject";function Mr(e,t){return Le(e.node)?(Lr(e.node),qr(e.node),!0):!1}h.processIfWorkletContextObject=Mr;function Le(e){return e.properties.some(t=>(0,v.isObjectProperty)(t)&&(0,v.isIdentifier)(t.key)&&t.key.name===h.contextObjectMarker)}h.isContextObject=Le;function qr(e){let t=(0,v.functionExpression)(null,[],(0,v.blockStatement)([(0,v.returnStatement)((0,v.cloneNode)(e))],[(0,v.directive)((0,v.directiveLiteral)("worklet"))]));e.properties.push((0,v.objectProperty)((0,v.identifier)(`${h.contextObjectMarker}Factory`),t))}function Lr(e){e.properties=e.properties.filter(t=>!((0,v.isObjectProperty)(t)&&(0,v.isIdentifier)(t.key)&&t.key.name===h.contextObjectMarker))}});var Be=f(T=>{"use strict";Object.defineProperty(T,"__esModule",{value:!0});T.isImplicitContextObject=T.processIfWorkletFile=void 0;var b=require("@babel/types"),ze=E(),Te=ue();function zr(e,t){return e.node.directives.some(r=>r.value.value==="worklet")?(e.node.directives=e.node.directives.filter(r=>r.value.value!=="worklet"),Tr(e),!0):!1}T.processIfWorkletFile=zr;function Tr(e){let t=e.get("body");Jr(e.node),t.forEach(r=>{let n=Nr(r);de(n)})}function Nr(e){return e.isExportNamedDeclaration()||e.isExportDefaultDeclaration()?e.get("declaration"):e}function de(e){(0,ze.isWorkletizableFunctionPath)(e)?(e.isArrowFunctionExpression()&&Gr(e.node),Ne(e.node.body)):(0,ze.isWorkletizableObjectPath)(e)?Ue(e)?Hr(e.node):Br(e):e.isVariableDeclaration()?Ur(e):e.isClassDeclaration()&&Vr(e.node.body)}function Ur(e){e.get("declarations").forEach(r=>{let n=r.get("init");n.isExpression()&&de(n)})}function Br(e){e.get("properties").forEach(r=>{if(r.isObjectMethod())Ne(r.node.body);else if(r.isObjectProperty()){let n=r.get("value");de(n)}})}function Gr(e){(0,b.isBlockStatement)(e.body)||(e.body=(0,b.blockStatement)([(0,b.returnStatement)(e.body)]))}function Ne(e){e.directives.some(t=>t.value.value==="worklet")||e.directives.push((0,b.directive)((0,b.directiveLiteral)("worklet")))}function Hr(e){e.properties.some(t=>(0,b.isObjectProperty)(t)&&(0,b.isIdentifier)(t.key)&&t.key.name===Te.contextObjectMarker)||e.properties.push((0,b.objectProperty)((0,b.identifier)(`${Te.contextObjectMarker}`),(0,b.booleanLiteral)(!0)))}function Ue(e){return e.get("properties").some(r=>r.isObjectMethod()?Zr(r):!1)}T.isImplicitContextObject=Ue;function Zr(e){let t=!1;return e.traverse({ThisExpression(r){t=!0,r.stop()}}),t}function Vr(e){e.body.push((0,b.classProperty)((0,b.identifier)("__workletClass"),(0,b.booleanLiteral)(!0)))}function Jr(e){let t=e.body,r=t.length,n=0;for(;n<r;){let i=t[n];if(!Xr(i)){n++;continue}let s=t.splice(n,1);t.push(...s),r--}}function Xr(e){return(0,b.isExpressionStatement)(e)&&(0,b.isAssignmentExpression)(e.expression)&&(0,b.isMemberExpression)(e.expression.left)&&(0,b.isIdentifier)(e.expression.left.object)&&e.expression.left.object.name==="exports"}});var Ge=f(re=>{"use strict";Object.defineProperty(re,"__esModule",{value:!0});re.processInlineStylesWarning=void 0;var d=require("@babel/types"),$r=D(),fe=require("assert");function Yr(e){return(0,d.callExpression)((0,d.arrowFunctionExpression)([],(0,d.blockStatement)([(0,d.expressionStatement)((0,d.callExpression)((0,d.memberExpression)((0,d.identifier)("console"),(0,d.identifier)("warn")),[(0,d.callExpression)((0,d.memberExpression)((0,d.callExpression)((0,d.identifier)("require"),[(0,d.stringLiteral)("react-native-reanimated")]),(0,d.identifier)("getUseOfValueInStyleWarning")),[])])),(0,d.returnStatement)(e.node)])),[])}function Kr(e){e.isMemberExpression()&&(0,d.isIdentifier)(e.node.property)&&e.node.property.name==="value"&&e.replaceWith(Yr(e))}function Qr(e){if((0,d.isArrayExpression)(e.node)){let t=e.get("elements");(0,fe.strict)(Array.isArray(t),"[Reanimated] `elements` should be an array.");for(let r of t)r.isObjectExpression()&&pe(r)}}function pe(e){let t=e.get("properties");for(let r of t)if(r.isObjectProperty()){let n=r.get("value");(0,d.isIdentifier)(r.node.key)&&r.node.key.name==="transform"?Qr(n):Kr(n)}}function en(e,t){if((0,$r.isRelease)()||t.opts.disableInlineStylesWarning||e.node.name.name!=="style"||!(0,d.isJSXExpressionContainer)(e.node.value))return;let r=e.get("value").get("expression");if((0,fe.strict)(!Array.isArray(r),"[Reanimated] `expression` should not be an array."),r.isArrayExpression()){let n=r.get("elements");(0,fe.strict)(Array.isArray(n),"[Reanimated] `elements` should be an array.");for(let i of n)i.isObjectExpression()&&pe(i)}else r.isObjectExpression()&&pe(r)}re.processInlineStylesWarning=en});var Ze=f(ne=>{"use strict";Object.defineProperty(ne,"__esModule",{value:!0});ne.substituteWebCallExpression=void 0;var He=require("@babel/types");function tn(e){let t=e.node.callee;if((0,He.isIdentifier)(t)){let r=t.name;(r==="isWeb"||r==="shouldBeUseWeb")&&e.replaceWith((0,He.booleanLiteral)(!0))}}ne.substituteWebCallExpression=tn});var $e=f(N=>{"use strict";var Ve=N&&N.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(N,"__esModule",{value:!0});N.processIfWorkletClass=void 0;var rn=Ve(require("@babel/generator")),nn=Ve(require("@babel/traverse")),a=require("@babel/types"),ie=require("assert"),on=E(),sn=D(),an=J(),Je="__workletClass";function ln(e,t){return!e.node.id||!bn(e.node.body)?!1:(mn(e.node.body),cn(e,t),!0)}N.processIfWorkletClass=ln;function cn(e,t){(0,ie.strict)(e.node.id);let r=e.node.id.name,n=un(e.node,t);yn(n),dn(n.program.body),fn(n.program.body,r),n.program.body.push((0,a.returnStatement)((0,a.identifier)(r)));let i=(0,a.functionExpression)(null,[],(0,a.blockStatement)([...n.program.body])),s=(0,a.callExpression)(i,[]);(0,sn.replaceWithFactoryCall)(e,r,s)}function un(e,t){let r=(0,rn.default)(e).code,n=(0,an.workletTransformSync)(r,{extraPlugins:["@babel/plugin-transform-class-properties","@babel/plugin-transform-classes","@babel/plugin-transform-unicode-regex"],filename:t.file.opts.filename,ast:!0,babelrc:!1,configFile:!1});return(0,ie.strict)(n&&n.ast),n.ast}function dn(e){e.forEach(t=>{if((0,a.isFunctionDeclaration)(t)){let r=(0,a.directive)((0,a.directiveLiteral)("worklet"));t.body.directives.push(r)}})}function fn(e,t){let r=t+on.workletClassFactorySuffix,n=pn(e,t),s=e[n].declarations[0].init,l=(0,a.functionDeclaration)((0,a.identifier)(r),[],(0,a.blockStatement)([(0,a.variableDeclaration)("const",[(0,a.variableDeclarator)((0,a.identifier)(t),s)]),(0,a.expressionStatement)((0,a.assignmentExpression)("=",(0,a.memberExpression)((0,a.identifier)(t),(0,a.identifier)(r)),(0,a.identifier)(r))),(0,a.returnStatement)((0,a.identifier)(t))],[(0,a.directive)((0,a.directiveLiteral)("worklet"))])),u=(0,a.variableDeclaration)("const",[(0,a.variableDeclarator)((0,a.identifier)(t),(0,a.callExpression)((0,a.identifier)(r),[]))]);e.splice(n,1,l,u)}function pn(e,t){let r=e.findIndex(n=>(0,a.isVariableDeclaration)(n)&&n.declarations.some(i=>(0,a.isIdentifier)(i.id)&&i.id.name===t));return(0,ie.strict)(r>=0),r}function bn(e){return e.body.some(t=>(0,a.isClassProperty)(t)&&(0,a.isIdentifier)(t.key)&&t.key.name===Je)}function mn(e){e.body=e.body.filter(t=>!(0,a.isClassProperty)(t)||!(0,a.isIdentifier)(t.key)||t.key.name!==Je)}function yn(e){let t=kn(e),r=gn(t),n=t.map(u=>u.index),i=r.map(u=>u.index),s=e.program.body,l=[...s];for(let u=0;u<t.length;u++){let p=i[u],k=n[u],_=l[p];s[k]=_}}function kn(e){let t=[];return(0,nn.default)(e,{Program:{enter:r=>{r.get("body").forEach((i,s)=>{var l;let u=i.getBindingIdentifiers();if(!i.isFunctionDeclaration()||!(!((l=i.node.id)===null||l===void 0)&&l.name))return;let p={name:i.node.id.name,index:s,dependencies:new Set};t.push(p),i.traverse({Identifier(k){_n(k,u,i)&&p.dependencies.add(k.node.name)}})})}}}),t}function gn(e){let t=[],r=new Set;for(let n of e)Xe(n,e,t,r);return t}function Xe(e,t,r,n){if(n.has(e.name))throw new Error("Cycle detected. This should never happen.");if(!r.find(i=>i.name===e.name)){n.add(e.name);for(let i of e.dependencies)if(!r.find(s=>s.name===i)){let s=t.find(l=>l.name===i);(0,ie.strict)(s),Xe(s,t,r,n)}r.push(e),n.delete(e.name)}}function _n(e,t,r){return e.isReferencedIdentifier()&&!(e.node.name in t)&&!r.scope.hasOwnBinding(e.node.name)&&r.scope.hasReference(e.node.name)}});Object.defineProperty(exports,"__esModule",{value:!0});var Ye=qe(),On=ue(),vn=Be(),hn=V(),En=Ge(),Sn=E(),In=D(),xn=Ze(),wn=$(),Wn=$e();module.exports=function(){function e(t){try{t()}catch(r){throw new Error(`[Reanimated] Babel plugin exception: ${r}`)}}return{pre(t){e(()=>{t.workletNumber=1,(0,hn.initializeGlobals)(),In.addCustomGlobals.call(this)})},visitor:{CallExpression:{enter(t,r){e(()=>{(0,Ye.processCalleesAutoworkletizableCallbacks)(t,r),r.opts.substituteWebPlatformChecks&&(0,xn.substituteWebCallExpression)(t)})}},[Sn.WorkletizableFunction]:{enter(t,r){e(()=>{(0,wn.processIfWithWorkletDirective)(t,r)||(0,Ye.processIfAutoworkletizableCallback)(t,r)})}},ObjectExpression:{enter(t,r){e(()=>{(0,On.processIfWorkletContextObject)(t,r)})}},ClassDeclaration:{enter(t,r){e(()=>{(0,Wn.processIfWorkletClass)(t,r)})}},Program:{enter(t,r){e(()=>{r.workletNumber=1,(0,vn.processIfWorkletFile)(t,r)})}},JSXAttribute:{enter(t,r){e(()=>(0,En.processInlineStylesWarning)(t,r))}}}}};
 //# sourceMappingURL=index.js.map
