+++ packages/react-native-reanimated/plugin/src/workletFactory.ts
@@ -78,6 +78,7 @@ export function makeWorkletFactory(
     '(' + (fun.isObjectMethod() ? 'function ' : '') + codeObject.code + '\n)';
 
   const transformed = workletTransformSync(codeObject.code, {
+    extraPlugins: plugins,
     filename: state.file.opts.filename,
     ast: true,
     babelrc: false,
@@ -453,3 +454,14 @@ function makeArrayFromCapturedBindings(
 
   return Array.from(closure.values());
 }
+
+const plugins = [
+  require.resolve('@babel/plugin-transform-shorthand-properties'),
+  require.resolve('@babel/plugin-transform-arrow-functions'),
+  require.resolve('@babel/plugin-transform-optional-chaining'),
+  require.resolve('@babel/plugin-transform-nullish-coalescing-operator'),
+  [
+    require.resolve('@babel/plugin-transform-template-literals'),
+    { loose: true },
+  ],
+];
